Fix duplicate uuid being generated
